Finish implementing support for Android scoped storage and misc tweaks/fixes
* Imagine: Add support for opening files with the Android storage access framework on OS 4.4+ and full directories on OS 5.0+
* Imagine: Remove MANAGE_EXTERNAL_STORAGE Android permission so OS 11+ only uses scoped storage per Google Play policy
* Imagine: Update FSPicker::OnSelectFileDelegate to pass the full path and display name of the selected file
* Imagine: Add URI path support to FSPicker
* Imagine: Show File Locations view when going up the root directory in FSPicker and integrate permission requests
* Imagine: Simplify JNI interface for the Android storage access framework
* Imagine: Add IG_forward() utility macro and clean up uses of std::forward()
* Imagine: Use forwarding references in DelegateFunc call functions to avoid copying non-trivial objects
* Imagine: Add ApplicationContext::storagePath(), forEachInDirectoryUri(), permissionIsRestricted()
* Imagine: Rename ApplicationContext::nearestRootPath() to rootPathInfo() and add support for URIs
* Imagine: Split out directory_iterator implementation into separate DirectoryStream class
* Imagine: Add new constexpr URI encode/decode utility functions
* Imagine: Clean up ordering of FS headers and allow constructing a directory_entry without a dirent
* Imagine: Don't check Android URL open intent with package manager since it's restricted on recent OS versions
* Imagine: Fix missing icon when adding an Android app shortcut
* EmuFramework: Remove redundant content open menu entries since all functionality is now in FSPicker
* EmuFramework: Use the directory name of a URI as the content directory if available
* EmuFramework: Pass path's display name to EmuSystem::createSystemWithMedia()
* EmuFramework: Don't add missing content to the recent list when loading config file
* EmuFramework: Don't restore saved file paths if storage permission is restricted (like on Android 11+) when loading config file
* EmuFramework: Default to an empty content search path so the user can select one via FSPicker File Locations view
* EmuFramework: Use ApplicationContext::storagePath() for fallback save path
* EmuFramework: Show an error message if opened content needs an associated directory but it was opened via a file URI
* EmuFramework: Remove redundant file existence checks in StateSlotView
* C64.emu, MD.emu, MSX.emu, NEO.emu, PCE.emu: Finish implementing URI file path support
* 2600.emu, MD.emu PCE.emu: Fix error loading ROMs from IO that doesn't report a size like pipes
* C64.emu: Update link to the VICE system files
* C64.emu: Clean up old AutostartPrgDiskImage on startup
* C64.emu & MSX.emu: Do a basic validation on the directory set for system/BIOS files
* MSX.emu: Optimize file system access when loading a ROM
* MSX.emu: Speed up state loading by caching opened zip file
* NEO.emu: Cache content directory filenames to speed up building game list in GameListView
* Various code cleanup

+++ imagine/include/imagine/gui/TextTableView.hh
@@ -23,13 +23,13 @@
 class TextTableView : public TableView
 {
 public:
-	TextTableView(IG::utf16String name, ViewAttachParams attach, unsigned itemsHint):
+	TextTableView(IG::utf16String name, ViewAttachParams attach, size_t itemsHint):
 		TableView{std::move(name), attach, textItem}
 	{
 		textItem.reserve(itemsHint);
 	}
 
-	TextTableView(ViewAttachParams attach, unsigned itemsHint);
+	TextTableView(ViewAttachParams attach, size_t itemsHint);
 	void appendItem(IG::utf16String name, TextMenuItem::SelectDelegate del);
 	void setItem(size_t idx, IG::utf16String name, TextMenuItem::SelectDelegate del);
 	TextMenuItem &item(size_t idx);
